Изменения в User.java и UserDao.java
