--- src/main/java/org/example/User.java
@@ -1,54 +1,0 @@
-package org.example;
-
-public class User {
-    private int id;
-    private String username;
-    private String password;
-    private String email;
-    private String phone;
-
-    // Конструкторы, геттеры и сеттеры
-
-    public User(int id, String username, String password, String email, String phone) {
-        this.id = id;
-        this.username = username;
-        this.password = password;
-        this.email = email;
-        this.phone = phone;
-    }
-
-    public int getId() {
-        return id;
-    }
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPhone() {
-        return phone;
-    }
-    public void setPhone(String phone) {
-        this.phone = phone;
-    }
-}
