+++ src/main/java/org/tps/autorization/User.java
@@ -1,0 +1,20 @@
+package org.tps.autorization;
+
+import lombok.Data;
+
+@Data
+public class User {
+    private int id;
+    private String username;
+    private String password;
+    private String email;
+    private String phone;
+
+    public User(int id, String username, String password, String email, String phone) {
+        this.id = id;
+        this.username = username;
+        this.password = password;
+        this.email = email;
+        this.phone = phone;
+    }
+}
